Float speed

+++ recorder.py
@@ -92,9 +92,9 @@ def master():
         if cmd[:3] == 'REC':
             selected = int(cmd[3:])
         elif cmd[:3] == 'RWD':
-            speed = -int(cmd[3:])
+            speed = -float(cmd[3:])
         elif cmd[:3] == 'FWD':
-            speed = int(cmd[3:])
+            speed = float(cmd[3:])
 
         # Get recording
         try:
@@ -114,11 +114,11 @@ def master():
         if cmd == 'PLAY' or cmd[:3] == 'REC':
             next_pos_r = pos_r + blocksize
         elif cmd[:3] == 'RWD':
-            next_pos_r = max(0,pos_r + speed * blocksize)
+            next_pos_r = max(0,pos_r + int(speed * blocksize))
             if next_pos_r == 0:
                 cmd = 'STOP'
         elif cmd[:3] == 'FWD':
-            next_pos_r = pos_r + speed * blocksize
+            next_pos_r = pos_r + int(speed * blocksize)
 
 def slave(index, filename):
     with sf.SoundFile(filename, 'r+') as f:
